scripts: resync all build scripts

+++ tools/config-hot-fix.sh
@@ -6,13 +6,16 @@ DIR=$PWD
 unset CC
 . "${DIR}/.CC"
 
-cd ${DIR}/KERNEL/
+if [ -f ${DIR}/KERNEL/Makefile ] ; then
+	cd ${DIR}/KERNEL/
 
-cp -v "${DIR}/patches/debian.config" .config
-cp -v "${DIR}/patches/beagle.config" beagle.config
-make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" olddefconfig
-ARCH=${KERNEL_ARCH} ./scripts/kconfig/merge_config.sh -m -r .config beagle.config
-make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" olddefconfig
-cp -v .config "${DIR}/patches/defconfig"
+	cp -v "${DIR}/patches/debian.config" .config
+	cp -v "${DIR}/patches/beagle.config" beagle.config
+	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" olddefconfig
+	ARCH=${KERNEL_ARCH} ./scripts/kconfig/merge_config.sh -m -r .config beagle.config
+	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" olddefconfig
+	cp -v .config "${DIR}/patches/defconfig"
+	rm beagle.config || true
 
-cd ${DIR}/
+	cd ${DIR}/
+fi
